+++ CHANGELOG.md
@@ -29,6 +29,9 @@ Also adds a new ALFPath class to replace alf path functions.
 - setup.py
 - one.alf.files; use one.alf.path instead
 - one.alf.io.remove_uuid_file
+- one.alf.io.remove_uuid_recursive
+- one.util.ensure_list; use iblutil.util.ensure_list instead
+- one.remote.globus.create_globus_client; use one.remote.globus.Globus class instead
 
 ## [2.11.1]